FW-5585 Remove Unused Site Data Endpoint (#864)
* remove site data API and tests

* redo site API deletions
