+++ firstvoices/backend/urls.py
@@ -9,7 +9,7 @@ from backend.views.custom_order_recalculate_views import (
     CustomOrderRecalculatePreviewView,
     CustomOrderRecalculateView,
 )
-from backend.views.data_views import MTDSitesDataViewSet, SitesDataViewSet
+from backend.views.data_views import MTDSitesDataViewSet
 from backend.views.dictionary_views import DictionaryViewSet
 from backend.views.gallery_views import GalleryViewSet
 from backend.views.games_views import WordsyViewSet
@@ -48,7 +48,6 @@ sites_router.register(r"audio", AudioViewSet, basename="audio")
 sites_router.register(r"categories", CategoryViewSet, basename="category")
 sites_router.register(r"characters", CharactersViewSet, basename="character")
 sites_router.register(r"contact-us", ContactUsView, basename="contact-us")
-sites_router.register(r"data", SitesDataViewSet, basename="data")
 sites_router.register(r"mtd-data", MTDSitesDataViewSet, basename="mtd-data")
 sites_router.register(r"dictionary", DictionaryViewSet, basename="dictionaryentry")
 sites_router.register(
